+++ src/main/java/com/artur114/bananalib/math/m2d/vec/Vec2DM.java
@@ -144,6 +144,6 @@ public class Vec2DM implements IVec2DM {
 
     @Override
     public String toString() {
-        return "(" + this.x() + ", " + this.y() + ")";
+        return "(" + ((float) this.x()) + ", " + ((float) this.y()) + ")";
     }
 }
